Replace Int type with I64 for consistent numeric naming
- Update TypeConstructor::Int to TypeConstructor::I64 in type system
- Change VeltranoType::int() constructor to VeltranoType::i64()
- Update parser to recognize I64 instead of Int type annotations
- Update all source files, tests, and examples to use I64
- Fix type generation to properly map I64 → i64 in Rust output
- Update MutRef<Int> references to MutRef<I64> throughout codebase

+++ src/rust_interop.rs
@@ -107,7 +107,7 @@ impl RustType {
             | RustType::ISize
             | RustType::U32
             | RustType::U64
-            | RustType::USize => Ok(VeltranoType::int()),
+            | RustType::USize => Ok(VeltranoType::i64()),
             RustType::Bool => Ok(VeltranoType::bool()),
             RustType::Unit => Ok(VeltranoType::unit()),
             RustType::Never => Ok(VeltranoType::nothing()),
